Updated from vgcore, vgandroid and vgios projects

+++ core/include/cmd/mglocal.h
@@ -12,7 +12,7 @@ struct MgView;
 //! 本地化文字的辅助类
 struct MgLocalized {
     static std::string getString(MgView* view, const char* name);
-    static int formatString(char *buffer, size_t size, MgView* view, const char *format, ...);
+    static int formatString(char *buffer, int size, MgView* view, const char *format, ...);
 };
 
 #endif // TOUCHVG_CORE_LOCALSTR_H
